Took their site and got it running on Django.
-As of right now only the three pages under "Students and Parents" tab are included in urls and views but the rest can be easily added through grunt work.
-changes to fruitful_minds/urls.py
-changes to main/views.py
-created an app 'main' to hold the content
-added STATIC_URL = '/static/' to settings.py. All static images are in fruitful_minds/static
-added 'main' and 'fruitful_minds' to INSTALLED_APPS in settings.py
-for the three functional pages, took their static image urls and replaced it with static template tag
-changed timezone to America/Los_Angeles
-created superuser on localhost:<port_number>/admin/ with username:admin password:password
What needs to be done now is the chunks of text/image/content on each page needs to be given a field in the database so the admin can easily edit them through Django.
Very straightforward grunt work to add the rest of the pages after doing so.
Let me know if you see any problems with my work/approach.